tests: stack-args: Add multi-arch support
For aarch64 and powerpc64, up to eight arguments can be
passed using registers which makes this test fail as none
of currently used arguments are on stack. This is fixed
by adding extra arguments to test function in order to
guarantee that the last two arguments are always on stack.

Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>