tests: intptrcast: Add multi-arch support
The stack pointer register, number of register arguments
and the stack frame offsets at which stack arguments are
available can vary across architectures. This adds support
for aarch64 and powerpc64 to this test and introduces
shell script based wrappers for the oneliners that make
sure that the correct register names and offsets are used.

Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>